--- .settings/org.eclipse.core.resources.prefs
@@ -1,6 +1,0 @@
-eclipse.preferences.version=1
-encoding//interface/auth/__init__.py=utf-8
-encoding//interface/auth/views.py=utf-8
-encoding//interface/index/__init__.py=utf-8
-encoding//interface/index/views.py=utf-8
-encoding//util/result/result.py=utf-8
