+++ interface/auth/views.py
@@ -9,28 +9,35 @@ create file : C:/Users/Raymond/git/testerlife/interface/auth/views.py
 create time :2016年11月1日
 '''
 
-from flask import render_template , request
+from flask import render_template , request, current_app
+from flask_restful import Resource
 
-from interface import login, login_user, login_required , current_user , logout_user
-from interface.auth import api , Resource, auth
-from interface.db import User
-from util.result.result import result
+from interface.util.result.result import result
+from . import api , auth
+from .. import login, login_user, login_required , current_user , logout_user
+from ..model import User
 
 
-@auth.route('/' , methods=['GET', 'POST'])
-def homepage():
-    return render_template('login.html')
 
 @login.user_loader
-def load_user(id):
-    return User.query.filter_by(id=int(id)).first()
+def load_user(user_id):
+    try:
+        user = User.query.get(int(user_id))
+        return user
+    except:
+        return None
+
+@auth.route('/' , methods=['GET', 'POST'])
+def index():
+    return render_template('login.html')
 
 @api.resource('/login')
-class Login(Resource):
+class login(Resource):
     def post(self):
         args = request.form
-        user = User.query.filter_by(username=args['username']).first()
-        if user and user.userpassword == args['password']:
+        current_app.logger.info(args)
+        user = User.query.filter_by(username=args.get('username')).first()
+        if user and user.userpassword == args.get('password'):
             login_user(user, True)
             return result.success()
         return result.error()
