+++ interface/config.py
@@ -1,4 +1,7 @@
 DEBUG = True
+
+SECRET_KEY = '\x12my\x0bVO\xeb\xf8\x18\x15\xc5_?\x91\xd7h\x06AC'
+# SQLALCHEMY
 SQLALCHEMY_DATABASE_URI = 'mysql+pymysql://root:autotest@127.0.0.1:3306/userInfo'
 SQLALCHEMY_POOL_SIZE = 50
 SQLALCHEMY_POOL_TIMEOUT = 60
