+++ run.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+from interface import app
+
+
+if __name__ == '__main__':
+    app.run(host='0.0.0.0', port=5000)
