--- util/__init__.py
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-# -- coding: utf-8 --
-'''
-File Name : util.__init__
-Description :
-Author : Raymond
-change Activity :
-create file : C:/Users/Raymond/git/testerlife/util/__init__.py
-create time :2016年11月1日
-'''
-
-"""
-    pass
-"""
